% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Italy}
\alias{Italy}
\title{Italian females}
\format{
A 5 column by 46 row data frame.
}
\usage{
data(Italy)
}
\description{
Italian females
}
\details{
This is a subset of the "Europe" data frame that includes only
Italian females.
Data is as follows:
\itemize{
 \item{Stat:}{  Stature in mm.}
 \item{Fem:}{  Maximum femur length in mm.}
 \item{Tib:}{  Maximum tibia length in mm.}
 \item{Hum:}{  Maximum humerus length in mm.}
 \item{Rad:}{  Maximum radius length in mm.}
}
}
\keyword{datasets}
