% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_run.R
\name{K_run}
\alias{K_run}
\title{Konigsberg & Jantz (2018) stature regressions}
\usage{
K_run(i.which = "Fem", which.dat = 2)
}
\arguments{
\item{which.dat}{Integer indicating which data to use from Ruff's European data. The default
default value is 2, which uses "Europe2" or any other integer which uses "Europe".
See the description of these two data objects.}

\item{w.which}{Character name indicating which long bone to use.  Possible
values are "Fem" for femur, "Tib" for tibia, "Hum" for humerus, and
"Rad" for radius.  Default: "Fem"}
}
\value{
Plot and other info.
}
\description{
Does plot and returns some counts
}
\details{
Uses the classical calibration method.
}
